requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
lxml==4.9.3
magic-filter==1.0.10
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
pypika-tortoise==0.1.6
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tortoise-orm==0.19.3
typing_extensions==4.7.1
tzdata==2023.3
yarl==1.9.2